move xarray object to new `.types` module
